moving getAllRegistrations to separate file
